requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
aiohttp>=3.9.0,<3.11
python-dotenv==1.0.1
pyrogram==2.0.106
TgCrypto==1.2.5
